Changed to double precision on CPU for water classes, also removed a lot of shader matrix multiplications.

+++ Source/Terrain/Water.h
@@ -3,7 +3,7 @@
 
 class Water : public Drawable {
 public:
-	Water(float radius, const glm::mat4 &translation, const glm::mat4 &scale, const glm::mat4 &rotation);
+	Water(double  radius, const glm::dmat4 &translation, const glm::dmat4 &scale, const glm::dmat4 &rotation);
 
 	Water() = delete;
 	Water(const Water&) = delete;
@@ -15,7 +15,7 @@ private:
 	class WaterMessage;
 
 	const unsigned int _base_resolution = 512;
-	float _water_level;
+	double _water_level;
 
 	const glm::mat4 _translation;
 	const glm::mat4 _scale;
